+++ backend/server.py
@@ -1,11 +1,18 @@
 from flask import Flask
 from flask_cors import CORS
+from flask_caching import Cache
 
 server = Flask(__name__)
+
 server.config.update(
     DEBUG=True,
     SECRET_KEY="TGSWORLDisAwesome",
+    CACHE_TYPE= "SimpleCache",  # Flask-Caching related configs
+    CACHE_DEFAULT_TIMEOUT= 300
 )
+
+cache = Cache(server)
+
 import db as dbInit
 
 
@@ -25,7 +32,6 @@ mailer=Mailer(cfg)
 if __name__ == '__main__':
     import endpoints.backoffice
     import endpoints.tickets
-    # TODO: import test endpoints only in debug mode
 
     if server.config.get('DEBUG'):
         import endpoints.test