+++ web/src/layouts/ProtectedRoute.jsx
@@ -4,19 +4,34 @@ import Cookies from 'js-cookie';
 import { checkAuthentication } from "../../../package/api/auth";
 import { Spinner } from "flowbite-react";
 import AdminLayout from "./AdminLayout";
+import Helmet from "react-helmet"
+import { getAdminPageTitle } from "../../../package/api/aAppData";
+
+
 
 export default function ProtectedRoute  ({
     redirectPath = '/admin/authenticate',
     children,
   })  {
-    const [authToken] = useState(Cookies.get("__tgssessiontoken"))
+    const [authToken] = useState(Cookies.get("__open-tickets-sessiontoken"))
     const [authenticated,setAuthenticated] =  useState(false)
     const [loading,setLoading] =  useState(true)
-  
+    
+    const [appTitle,setAppTitle]= useState(Cookies.get("__open-tickets-app-title__cache"))
+    
   
     useEffect(()=>{
       async function callMeBaby(){
+        if(!appTitle){
+          let title=(await getAdminPageTitle()).title
+          Cookies.set("__open-tickets-app-title__cache",title)
+          setAppTitle(title)
+        
+        }
         let r= await checkAuthentication(authToken)
+
+
+
         setAuthenticated(r)
         setLoading(false)
       }
@@ -35,7 +50,10 @@ export default function ProtectedRoute  ({
         </div>
       </>)
     }
-    return <AdminLayout>
+    return <AdminLayout appTitle={appTitle}  >
+      <div>
+            <Helmet title={appTitle}/>
+      </div>
       {children}
     </AdminLayout>;
   };